LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/reporting/reporting-isolated-across-navigations.https.sub-expected.txt

Bug test main frame



FAIL Reports should be sent to the correct endpoints assert_equals: First endpoint should receive two reports expected 2 but got 1
